[ALPS04589232] wlan: Support WPA3-5.2.1
Supplicant uses AES-CCMP as the key_mgmt and OPEN
as the auth_type to reconnect to a AP if Supplicant
once associated with this AP before. Typically, an OPEN
AP doesnt have a key_mgmt, but Supplicant uses OPEN
to skip another SAE exchange and uses AES-CCMP to fill
PMKID in assoc. That is, driver has to support this new
scenario.

bug fix: remove driver-filled RSN IE.

MTK-Commit-Id: a126d5f91e98725e96b51a1e3cd0297b78e773f2

Change-Id: I69cd2d611b39d59128468ddc1611198562961c23
CR-Id: ALPS04589232
Feature: [Module]Wi-Fi Driver
Signed-off-by: Wayne-TW Chen <wayne-tw.chen@mediatek.com>