[ALPS04707054] wlan: Dont RESET if recv unprotected mgmt
[Detail]
CFG_SUPPORT_802_11W was turned on because of WPA3. In SAE-5.2.4, AP sent
an encrypted deauth 3 to STAUT and then it sent a unprotected deauth 8
to STAUT.
NIC uses rsnCheckRxMgmt() to check if the mgmt frame should be protected.
If it should but it doesnt, NIC then triggers chip reset.

[Solution]
We shall ignore unprotected mgmt frame from a PMF AP. Remove chip reset
at this scenario.

MTK-Commit-Id: 58414f6a0d975596616eb1cf80a09cc208eb2c22

Change-Id: I44fc8e32fd938b9911166d176d88cba2cc8de521
CR-Id: ALPS04707054
Feature: [Module]Wi-Fi Driver
Signed-off-by: Wayne-TW Chen <wayne-tw.chen@mediatek.com>