[ALPS04794888] wlan: Skip SAQuery if unassociated
In 802.11w, sta has to continuously send SA query to the AP if it does
not receive any SA response from the AP. In our design:
1. we keep sending SA query every 200ms if we cannot receive SA response
2. if 5 SA responses are all missing, then we trigger disconnect
That is, we keep sending SA queries no matter what the connection
state is.
If we disconnect with the AP and another 200ms elapses, we will still
try to send another SA query. This SA query causes kernel panic since
all the information about the previous connection is freed.
We have to add another terminus for disconnection.

MTK-Commit-Id: 0003973ba7a3f978821cb7fff57e33a0c653be64

Change-Id: I89faa9cc9f639b64ee2c9d69f53aecae18df02c6
CR-Id: ALPS04794888
Feature: [Module]Wi-Fi Driver
Signed-off-by: Wayne-TW Chen <wayne-tw.chen@mediatek.com>